requirements.txt
requests
python-dotenv
discord-webhook
